Support use with linkers/setups needing rpath-link (#8)
Discovered in testing of integration of G4VG in apt-sim/AdePT#289,
targets linking to installed G4VG may fail to link due to linker
not having needed information on location of deps-of-deps of G4VG.
This happens in build environments where

1. The installed dependencies (here, Geant4, VecGeom) do not have
   their rpath set correctly.
2. They are otherwise in a location the linker is not aware of,
   e.g. via ldconfig of LD_LIBRARY_PATH etc.

The link error warns about this and suggests adding the path through,
e.g. `-rpath-link`. CMake will always add this flag/path correctly,
but only if it knows about the dependencies (even if private).

The G4VG library only depends on the static Celeritas::geocel target,
which does not forward its dependencies to Geant4/Vecgeom and so
it does not set `-rpath-link` for these.

Workaround this issue by explicitly making used Geant4 targets
PRIVATE deps of the g4vg target. Refind Geant4/VecGeom dependencies
when finding G4VG so these are resolved and available if the linker
requires it.

Add smoke test for linking that simply tries to link only the G4VG
target. Confirmed to reproduce problem observed in AdePT, with
the fixes above resolving it.